Automate the generation & future updates of the release date markdown files (#2)
Automate the generation & future updates of the release date markdown files.

This change brings:
1) Automation to generate the release date markdown files
2) A GitHub Workflow to run aforementioned automation on a schedule (hourly), committing any changes as needed

This change adds support for the below IDEs:
- AppCode
- CLion
- GoLand
- IntelliJ IDEA Ultimate
- PyCharm Professional Edition
- PhpStorm
- ReSharper C++
- Rider
- RubyMine
- ReSharper
- WebStorm

(See the FAQ entry in the `README.md` file for why IDEA & PyCharm Community are not included; tl;dr, they are near-identical to the respective paid offerings.)

A JetBrains provided API to get all the necessary release date information and generate the markdown tables accordingly. The GitHub workflow automates the generation of these files (currently, every hour), and if there are changes to any of the markdown files in the `ide/` directory, they are automatically committed. Additionally, a fail-safe is provided to ensure the changes to the markdown files are additions of release dates, and not accidental deletions.